+++ src/main/java/net/meander/subtlyd/sounds/SoundEventsSD.java
@@ -15,6 +15,7 @@ public class SoundEventsSD {
     public static final SoundEvent SNOW_BRICK_PLACE = register("block.snow_bricks.place");
     public static final SoundEvent SNOW_BRICK_STEP = register("block.snow_bricks.step");
     public static final SoundEvent WITHER_SKELETONS_SUMMONED = register("entity.wither_skeleton.summon");
+    public static final SoundEvent EVOKER_FANGS_APPEAR = register("entity.evoker_fangs.appear");
     public static final SoundEvent STICK_LIGHT = register("item.stick.light");
 
     public static void registration() { }